+++ src/main/java/terramine/datagen/ModItemsBlocksModels.java
@@ -103,6 +103,7 @@ public class ModItemsBlocksModels extends FabricModelProvider {
         }
         generateFlatItem(itemModelGenerators, ModItems.MAGIC_MIRROR, ModelTemplates.FLAT_ITEM, "misc");
         generateFlatItem(itemModelGenerators, ModItems.SHIMMER_BUCKET, ModelTemplates.FLAT_ITEM, "misc");
+        generateFlatWithHeldModelItem(itemModelGenerators, ModItems.CLENTAMINATOR, "tools/clentaminator");
 
         // Custom Blocks
         generateBlockItem(itemModelGenerators, ModBlocks.CORRUPTED_SNOW_LAYER.ITEM, "corrupted_snow_height2");